requirements.txt
python-telegram-bot==20.3
yt-dlp
python-dotenv
gallery-dl
